requirements.txt
discord.py
requests
python-decouple
youtube_dl
asyncio
PyNaCl